changed data type for positions from int to size_t

+++ src/lox/Scanner.hpp
@@ -15,8 +15,8 @@ class Scanner
   private:
     std::string m_source;
     std::vector<Token> m_tokens;
-    int m_start = 0;
-    int m_current = 0;
+    size_t m_start = 0;
+    size_t m_current = 0;
     int m_line = 1;
     std::unordered_map<std::string, TokenType> m_reserved;
   public:
